chatbot/requirements.txt
# python=3.8.16

google-cloud-bigquery==3.20.1
google-cloud-storage
langchain==0.1.15
langchain-openai==0.0.8
sqlalchemy==1.4.52
pybigquery
faiss-cpu==1.8.0

pandas
gradio==4.21.0
plotly==5.20.0
chromadb==0.4.19
docx2txt==0.8
unstructured==0.5.6
pypdf==4.1.0
gunicorn==21.2.0
Jinja2==3.1.3

statsmodels